pyproject.toml
[project]
name = "vedgyproject"
version = "0.1.0"
description = "Open-source vegan housing platform"
requires-python = ">=3.13"
dependencies = [
    "django==5.2.6",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.1.1",
    "pillow==11.3.0",
    "boto3==1.40.50",
    "gunicorn==23.0.0",
    "django-ratelimit==4.1.0",
    "b2sdk==2.10.0",
    "whitenoise==6.11.0",
    "dj-database-url==3.0.1",
    "pillow-heif==1.1.1",
    "pydantic==2.12.0",
    "django-ninja>=1.4,<2.0",
    "django-ninja-jwt>=5.3,<6.0",
    "django-cors-headers>=4.6,<5.0",
]

[dependency-groups]
dev = [
    "pytest==8.4.2",
    "pytest-django==4.11.0",
    "autoflake",
    "black",
    "isort",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--verbose --strict-markers --tb=short --reuse-db"
markers = ["django_db: Mark test as requiring database access"]
